+++ apps/sam_l21_xpro/saml21_ble_fitness_tracker/firmware/src/config/sam_l21_xpro/peripheral/pm/plib_pm.c
@@ -152,13 +152,13 @@ bool PM_ConfigurePerformanceLevel(PLCFG_PLSEL plsel)
     /* Write the value only if Performance Level Disable is not set */
     if ((PM_REGS->PM_PLCFG & PM_PLCFG_PLDIS_Msk) == 0U)
     {
-        if((PM_REGS->PM_PLCFG & PM_PLCFG_PLSEL_Msk) != plsel)
+        if((PM_REGS->PM_PLCFG & PM_PLCFG_PLSEL_Msk) != (uint8_t)plsel)
         {
             /* Clear INTFLAG.PLRDY */
             PM_REGS->PM_INTFLAG |= (uint8_t)PM_INTENCLR_PLRDY_Msk;
 
             /* Write PLSEL bits */
-            PM_REGS->PM_PLCFG  = plsel;
+            PM_REGS->PM_PLCFG  = (uint8_t)plsel;
 
             while((PM_REGS->PM_INTFLAG & PM_INTFLAG_PLRDY_Msk) == 0U)
             {
